Project_Companion_AI/requirements.txt
# requirements.txt
chainlit
google-generativeai
python-dotenv
SpeechRecognition
PyAudio # Required by SpeechRecognition for microphone input
pyttsx3